requirements.txt
streamlit
yfinance
pandas
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
google-generativeai
pypdf
requests
torchvision
python-dotenv